+++ src/Classes/Books/Ebook.java
@@ -6,7 +6,7 @@ public class Ebook extends BaseBook{
     private String fileType;
 
 
-    public Ebook(String ISBN , String title , int year , int price , String fileType) {
+    public Ebook(String ISBN , String title , int year , double price , String fileType) {
 
         super(ISBN , title , year , price);
         this.fileType = fileType;
